# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "User API",
    "description": "An API for managing users in a database",
    "version": "1.0.0"
  },
  "paths": {
...
    "/relations": {
      "get": {
        "tags": ["relations"],
        "summary": "Get all relations",
        "description": "Get all relations from the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfully retrieved all relations",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Relations"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve relations from the database"
          }
        }
      },
      "post": {
        "tags": [
          "relations"
        ],
        "summary": "Insert new relation",
        "description": "Insert a new relation into the database",
        "parameters": [
        {
          "name": "sender_id",
          "in": "formData",
          "required": true,
          "type": "integer"
        },
        {
          "name": "receiver_id",
          "in": "formData",
          "required": true,
          "type": "integer"
        },
        {
          "name": "type",
          "in": "formData",
          "required": true,
          "type": "string"
        }
        ],
        "responses": {
          "201": {
            "description": "Successfully inserted user into the database"
          },
          "500": {
            "description": "Failed to insert user into the database"
          }
        }
      },
      "delete": {
        "tags": ["relations"],
        "summary": "Delete a relation",
        "description": "Delete a relation",
        "parameters": [
          {
            "name": "sender_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "receiver_id",
            "in": "formData",
            "required": true,
            "type": "integer"
          },
          {
            "name": "type",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the relation",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Relations"
              }
            }
          },
          "500": {
            "description": "Failed to delete relation from the database"
          }
        }
      }
    },
    "/relations/{sender_id}": {
      "get": {
        "tags": ["relations"],
        "summary": "Get all blocked by sender_id",
        "description": "Get all blocked by sender_id from the database",
        "parameters": [
        {
          "name": "sender_id",
          "in": "path",
          "required": true,
          "type": "integer"
        }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all blocked by sender_id",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Relations"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve relations from the database"
          }
        }
      }
    },
    "/relations/{sender_id}/blocked": {
      "get": {
        "tags": ["relations"],
        "summary": "Get users blocked by sender_id",
        "description": "Get users blocked by sender_id from the database",
        "parameters": [
          {
          "name": "sender_id",
          "in": "path",
          "required": true,
          "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all blocked by sender_id",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Relations"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve users blocked by sender_id from the database"
          }
        }
      }
    },
    "/relations/{sender_id}/liked": {
      "get": {
        "tags": ["relations"],
        "summary": "Get all users liked by sender_id",
        "description": "Get all users liked by sender_id from the database",
        "parameters": [
          {
          "name": "sender_id",
          "in": "path",
          "required": true,
          "type": "integer"
        }
      ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all users liked by sender_id",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Relations"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve users liked by sender_id from the database"
          }
        }
      }
    },
    "/relations/{sender_id}/matched": {
      "get": {
        "tags": ["relations"],
        "summary": "Get all users matched by sender_id",
        "description": "Get all users matched by sender_id from the database",
        "parameters": [
          {
          "name": "sender_id",
          "in": "path",
          "required": true,
          "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all users matched by sender_id",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Relations"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve users matched by sender_id from the database"
          }
        }
      }
    },
    "/relations/type/{sender_id}/{receiver_id}": {
      "get": {
        "tags": ["relations"],
        "summary": "Get relation type between from sender_id to receiver_id",
        "description": "Get type of relation from sender_id to receiver_id from the database",
        "parameters": [
          {
          "name": "sender_id",
          "in": "path",
          "required": true,
          "type": "integer"
          },
          {"name": "receiver_id",
          "in": "path",
          "required": true,
          "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved type of relation from sender_id to receiver_id",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Relations"
              }
            }
          },
          "500": {
            "description": "Failed to retrieve type of relation from sender_id to receiver_id"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Relations": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "description": "The unique identifier for the conversation"
        },
        "sender_id": {
          "type": "integer",
          "description": "The id of the user who sent the first message in the conversation"
        },
        "receiver_id": {
          "type": "integer",
          "description": "The id of the user who received the first message in the conversation"
        },
        "type": {
          "type": "string",
          "description": "The string that tells the relationship type, can be block, like or match"
        }
      }
    },
...
  }
}